Refactored some code and implemented an option for Async saving
Currently the Async option is just a static boolean variable (I'm
confident of it's effectiveness however testing should be done)
